Move client tags services to src/api
- Remove duplicated code
